Creates EAS build related files
Warning for future:

When setting up react-navigation, be
sure to follow the installation instructions
on the website! You have to install some
extra packages  (or else you'll get that RNSScreen error).

What's sneaky is that if you fail to do this, all local builds
still work. It's only the EAS builds that'll fail.