Add exercise ordering
Consider the scenario where a routine has 15 (ordered)
exercises. If I move the last to the first, then
the 'order' or 'index' of every single one of those
exercises changes (15 -> 1, 1 -> 2, 2-> 3, ...).

Under our current implementation, this requires 15
consecutive asynchronous DB update calls.

There is the approach of 'normalizing' by keeping the
indices spaced (e.g., if we have indices 10, 20, 30, ...
then it suffices to do 150 -> 9), but
    a) It is substantially more complicated
    b) Requires periodic 'renormalization'.

We should switch to the normalized approach
only after stress-testing with a reasonable input size.