Refactor timer state management
Turns out we needed to make the 'handleFinish' functionality
more sophisticated.

We create a general function for handling
changes to the timer -- both forward and backward.

This way forward skips, backward skips, and
organic timer completions all use the same function.

To remind ourselves of what needs to be done
in each of these cases:
    - Ring needs to be refilled
    - Duration needs to be reset
    - Elapsed time needs to be corrected

TODO:
    Need to sync all three components:
        progress,
        numerical timer,
        ring timer