PARTIAL COMMIT
Turns out we need to make the 'handleFinish' functionality
more sophisticated.

We should probably make a general function for handling
changes to the timer -- both forward and backward.

This way forward skips, backward skips, and
organic timer completions all use the same function.
To remind ourselves of what needs to be done
in each of these cases:
    - Ring needs to be refilled
    - Duration needs to be reset
    - (not now) Elapsed time needs to be corrected